requirements.txt
fastapi
uvicorn
python-multipart
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma
langchain-openai
chromadb
sentence-transformers
GitPython
python-dotenv
tiktoken
pydantic
